fix(frontend): keep dashboard tab icons out of unit tests
